Move agcontext editor to component within search/
