--- form/docker-compose.yml
@@ -1,16 +1,0 @@
-version: "3"
-services:
-  app:
-    tty: true
-    stdin_open: true # without this node doesn't start
-  frontend:
-    build: ./App
-    command: ["yarn", "start"]
-    volumes: 
-      - ./App
-      - ./node_modules
-    ports:
-      - "3000:3000"
-
-volumes:
-  node-modules:
