+++ search/package.json
@@ -12,9 +12,14 @@
     "node-forge": "^0.10.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-jsonschema-form": "^1.8.1",
+    "react-jsonschema-form-manager": "^0.4.7",
+    "react-jsonschema-form-material-ui": "^2.2.1",
     "react-reactive-form": "^1.0.32",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3",
     "redux": "^2.0.0",
+    "rjsf-material-ui": "^0.3.9",
     "typescript": "^2.8.0"
   },
   "scripts": {
