+++ search/src/AgContextForm.css
@@ -1,19 +1,19 @@
-.App {
+.AgContextForm {
   text-align: center;
 }
 
-.App-logo {
+.AgContextForm-logo {
   height: 40vmin;
   pointer-events: none;
 }
 
 @media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
+  .AgContextForm-logo {
+    animation: AgContextForm-logo-spin infinite 20s linear;
   }
 }
 
-.App-header {
+.AgContextForm-header {
   background-color: #282c34;
   min-height: 100vh;
   display: flex;
@@ -24,11 +24,11 @@
   color: white;
 }
 
-.App-link {
+.AgContextForm-link {
   color: #61dafb;
 }
 
-@keyframes App-logo-spin {
+@keyframes AgContextForm-logo-spin {
   from {
     transform: rotate(0deg);
   }
